Fix bug in postgres conf template
The last line of the template needs to return a successful error code, or the whole thing breaks. Even though the test block catches the error, it still sets $?, which seems to make bash unhappy in the parent shell.

Fun times. The "noop" colon seems to make it happy again.