Add "safemode", allows running with no conf
Previously, haxby would fail to run if a config were not available. This
made it a bit hard to use, as even the config builder would not work.

Modes can now be marked as "safe" to allow them to run at all times.
Configuration is minimal at this point, so they should be read-only.

Fixes #8