Switched shell escape code from \e to \033 for Mac terminal compatibility.
Depending on the shell implementation, escape can vary. Which is a nice
way to say that OsX's basic terminal seems not to support \e as an
escape sequence for shell colors. However, it does obey it in octal:
\033. Just like VMS did.

The logic is: 27 = 033 = 0x1b = ^[ = \e
All are synonyms of Escape. (ASCII 27, octal 33, hex 1b, Ctrl-[ and \e)

In some cases they are shell interpreted symbols, and in some cases they
are literal escape characters. \033 and 0x1b tend to be handled as the
latter, while \e tends to be interpreted as the former. And that
tendency is why \033 seems to work on more various unix systems than \e.

Closes-Bug: None
Implements: Shell colors on native Osx terminal
Change-Id: